resources/jsondatetime/jsondatetime/jsondatetime.py
import json
import datetime
from datetime import datetime
import dateutil.parser

try:
    string_types = basestring  # Python 2
except NameError:
    string_types = str  # Python 3

DEFAULT_DATE_FORMAT = '%a, %d %b %Y %H:%M:%S UTC'
DEFAULT_ARGUMENT = "datetime_format"

class DatetimeJSONEncoder(json.JSONEncoder):

    def default(self, obj):
        if isinstance(obj, datetime.datetime) or isinstance(obj, datetime.date):
            return obj.isoformat()
        else:
            return json.JSONEncoder.default(obj)

# I guess this thing only works when its not default? I think I only call load with it in that case
# but when I call dump with it (ie call dump when the line below is uncommented) i get the dreaded
# TypeError: encode() missing 1 required positional argument: 'o'

# json._default_encoder = DatetimeJSONEncoder


def dumps(obj, skipkeys=False, ensure_ascii=True, check_circular=True,
          allow_nan=True, cls=None, indent=None, separators=None,
          encoding='utf-8', default=None, sort_keys=False, **kw):
    return json.dumps(obj, skipkeys=skipkeys, ensure_ascii=ensure_ascii,
                      check_circular=check_circular, allow_nan=allow_nan,
                      cls=cls, indent=indent, separators=None, encoding=encoding,
                      default=default, sort_keys=sort_keys, **kw)

def loads(s, **kwargs):

    source = json.loads(s, **kwargs)

    return iteritems(source)

def iteritems(source):

    for k, v in source.items():
        if isinstance(v, list):
            for a in v:
                iteritems(a)
        elif isinstance(v, dict):
            iteritems(v)
        elif isinstance(v, string_types):
            try:
                # only acceptable format:
                # "2021-11-17T01:00:50.205Z"
                source[k] = datetime.strptime(v, "%Y-%m-%dT%H:%M:%S.%fZ")
            except:
                pass

    return source


